requirements.txt
discord.py
wheel
ffmpeg
youtube_dl
PyNaCl
requests
pygithub